Add support for CUE automation
