+++ requirements_test_all.txt
@@ -26,7 +26,7 @@ RtmAPI==0.7.2
 YesssSMS==0.4.1
 
 # homeassistant.components.abode
-abodepy==0.17.0
+abodepy==0.18.0
 
 # homeassistant.components.androidtv
 adb-shell==0.1.1
